git_sync

Module cli

Structs

AppArgs

git_sync is an application for managing multiple github repositories at once.

BackupRepoCommand

Backup repositories

CheckRepoCommand

Check repositories for various conditions

ClosePRCommand

Close a PR for a repository.

CompareTagCommand

Options for the ‘comparison’ tag command

CreateBranchCommand

Create a new branch.

CreatePRCommand

Open a PR to merge head into branch

CreateReleaseCommand

Create a release for a repository.

CreateTagCommand

Options for the ‘create’ tag command

DeleteBranchCommand

Delete a branch from a repository

DeleteTagCommand

Defines the arguments for the ‘delete’ tag subcommand

MergePRCommand

Merge a PR.

PruneBackupCommand

SyncRepoCommand

Sync a repository. If the repository is already up to date, no error is reported.

SyncTagCommand

Sync tags for a forked repository with its parent

Enums

BackupCommand

Define all the valid commands for acting on backups

BackupDestination

BranchCommand

Define all the valid commands for acting on branches

Command

The top-level command enum for the CLI

MakeLatest

Valid options that can be passed to make_latest in the github api.

PRCommand

Define all the valid commands for acting on pull requests

ReleaseCommand

Define all the valid commands for actiong on releases

RepoCommand

Define all the valid commands for acting on repositories

RepositoryType

Valid options for repository_type for cli arguments

TagCommand

Define all the valid commands for acting on Tags

Functions

cli

dir_exists 🔒

Check that the directory passed is a valid and existing directory

parse_args

Parse the command line arguments and validate them.

validate_jobs 🔒

Validate that the maximum number of paralllel jobs is between 1 and 64. Strictly speaking, there isn’t a reason that this couldn’t be higher, but there isn’t much point in allowing more jobs than cpu cores available