git_sync

Module github

Modules

backup

Backup repositories

branch

Managa branches for repositories

check

Run various checks and sanity checks for a repository

client

The top level definition for the Github client

fork

Manages forked repositories, particularly updating the fork to match the upstream repository

match_args

pr

Used to manage pull requests for libraries

release

Manage releases

tag

Manage tags