git_sync::github::client

Struct GithubClient

pub struct GithubClient {
    pub octocrab: Octocrab,
    pub semaphore: Arc<Semaphore>,
    pub webhook_url: String,
    pub is_tty: bool,
    pub output: Arc<OnceCell<OutputMode>>,
    slack_messages: Arc<Mutex<Vec<String>>>,
    slack_errors: Arc<Mutex<Vec<String>>>,
}

Github api entry point

Fields
octocrab: Octocrab

Octocrab client. This can be trivially cloned

semaphore: Arc<Semaphore>

A semaphore to control the maximum number of jobs that can be run in parallel

webhook_url: String

An optional webook for posting to slack, if that feature is enabled

is_tty: bool

Defines whether or not we’re running in interactive mode

output: Arc<OnceCell<OutputMode>>

Where output should go. This can only be written to Once

slack_messages: Arc<Mutex<Vec<String>>>

A message that gets sent to slack at the end of all processing, if it has any contents. This is thread safe.

slack_errors: Arc<Mutex<Vec<String>>>

An error message that will get sent to slack at the end of all processing, if it has any contents. This field is also thread safe.

Implementations

impl GithubClient

pub async fn backup_repo( &self, url: String, path: &Path, ) -> Result<(), GitError>

Backup a single repository into the specified folder

pub async fn backup_all_repos( &self, repositories: Vec<String>, path: &Path, ) -> Result<(), GitError>

Backup all configured repositories into the specified folder. This can take a long time, dependiing on the number of repositories and their sizes

pub async fn prune_backup( &self, path: &Path, since: Option<String>, ) -> Result<(), GitError>

This is a way to clean up stale references in a configured backed up repository. This is still under development, so don’t rely on it too much yet.

pub async fn backup_to_s3( &self, file_path: &Path, bucket: &str, key: &str, ) -> Result<(), GitError>

Upload a backup to s3. This requires your aws credentials to be configured in the environment. For small enough files, it will do a single part file upload but if the file is bigger than PART_SIZE, it will instead do a parallel multipart upload.

pub async fn backup_all_to_s3( &self, path: &Path, bucket: &str, ) -> Result<(), GitError>

Backuup all repositories to S3

impl GithubClient

pub async fn get_branch_sha( &self, url: &str, branch: &str, ) -> Result<String, GitError>

Get the most recent commit of a branch, so we can use that to create and delete it

pub async fn create_branch( &self, url: &str, base_branch: &str, new_branch: &str, ) -> Result<(), GitError>

Create a branch from some base branch in a repository

pub async fn create_all_branches( &self, base_branch: &str, new_branch: &str, repositories: &Vec<String>, ) -> Result<(), GitError>

Create the passed branch for each repository provided

pub async fn delete_branch( &self, url: &str, branch: &str, ) -> Result<(), GitError>

Delete a branch from a repository

pub async fn delete_all_branches( &self, branch: &str, repositories: &Vec<String>, ) -> Result<(), GitError>

Delete the specified branch for each configured repository

impl GithubClient

pub async fn check_repository( &self, url: &str, blacklist: HashSet<String>, checks: &RepoChecks, ) -> Result<Checks, GitError>

Get branches that are older than a certain number of days A blacklist should be passed to ignore certain branches, since some of them are static and will never change, nor should they be deleted.

pub fn display_check_results( header: Vec<String>, rows: Vec<Vec<String>>, rules: &[BranchProtectionRule], license: Option<&LicenseInfo>, repo: &str, )

Display the results of the various checks into a nice table

pub fn display_all_check_results( header: Vec<String>, rows: Vec<Vec<String>>, rules: &[BranchProtectionRule], license: Option<&LicenseInfo>, title: &str, repo: &str, )

TODO: This function is currently not used, but could be useful in the future

pub async fn check_all_repositories( &self, repositories: Vec<String>, checks: &RepoChecks, blacklist: &HashSet<String>, ) -> Result<Vec<Checks>, GitError>

Run the selected checks against all configured repositories

impl GithubClient

pub fn new( github_token: &str, config: &Config, max_jobs: usize, ) -> Result<Self, GitError>

pub async fn get_parent_repo(&self, url: &str) -> Result<RepoInfo, GitError>

Get the parent repository of a github repository.

pub async fn get_rate_limit(&self) -> Result<(), GitError>

Get the number of api calls left at the moment. Generally, the maximum number is 5000 in one hour.

pub async fn get_graphql_limit(&self) -> Result<(), GitError>

Get the number of graphql api calls left at the moment. Generally, the maximum number is 5000. This is tracked separately from the rest api limits.

pub async fn append_slack_message(&self, msg: String)

Appends success messages to be sent to slack at the end. The message will only be sent if slack integration is enabled and the webhook url is set.

pub async fn append_slack_error(&self, msg: String)

Append an error message to be sent to slack at the end. The message will only be sent if slack integartion is enabled and the webhhook url is set.

pub async fn slack_message(&self)

This can be used to send the contents of slack_messages and slack_errors to slack in one batch.

impl GithubClient

pub async fn sync_fork(&self, url: &str) -> Result<(), GitError>

Sync a single repository with its parent repository.

pub async fn sync_all_forks( &self, repositories: Vec<String>, ) -> Result<(), GitError>

Sync all configured repositories. Only repositories that have a parent repository should be passed to this function

impl GithubClient

pub async fn create_pr(&self, opts: &CreatePrOptions) -> Result<u64, GitError>

Create a pull request for a specific repository

pub async fn create_all_prs( &self, opts: &CreatePrOptions, merge_opts: Option<MergePrOptions>, repositories: Vec<String>, ) -> Result<HashMap<String, u64>, GitError>

Create a pull request for all configured repositories, and optionally merge them automatically, if possible

pub async fn merge_pr(&self, opts: &MergePrOptions) -> Result<(), GitError>

Merge a pull request. This will only work if there are no merge conflicts in the pull request.

impl GithubClient

pub async fn generate_release_notes( &self, url: &str, tag: &str, previous_tag: &str, ) -> Result<ReleaseNotes, GitError>

Generate release notes for a particular releaese. It grabs all the commits present in tag that are newer than the latest commit in previous_tag. This needs to be cleaned up, it is a bit of a mess right now.

pub async fn create_release( &self, url: &str, current_tag: &str, previous_tag: &str, release_name: Option<&str>, ) -> Result<(), GitError>

Create a new release for a specific repository. Release notes will also be generated.

pub async fn create_all_releases( &self, current_tag: &str, previous_tag: &str, release_name: Option<&str>, repositories: Vec<String>, ) -> Result<(), GitError>

Create a new release for each of the specified repositories.

impl GithubClient

pub async fn get_tags( &self, url: &str, ) -> Result<(IndexSet<TagInfo>, HashSet<String>), GitError>

This can be used to fetch tags in a more api-call efficient way than using the rest api. It does mean we have to manually query the graphql endpoint and manually parse the json output, rather than having it done for us by octocrab. We can’t actually get all the information required for an annotated tag, but we can use it to distinguish between lightweight and annotated tags. If we don’t get any annotated tags, we can skip the fairly slow git clone and push process

IndexSet is an implementation of an orderered Set.

pub async fn compare_tags( &self, url: &str, parent: &RepoInfo, ) -> Result<Comparison, GitError>

pub async fn diff_tags(&self, url: &str) -> Result<Comparison, GitError>

Get a diff of tags between a single forked repository and its parent repository.

pub async fn diff_all_tags( &self, repositories: Vec<String>, ) -> Result<(), GitError>

Get a diff of all configured repositories tags, compared against their parent.

pub async fn sync_tags( &self, url: &str, process_annotated: bool, ) -> Result<(), GitError>

Sync many tags asynchronously. We can use the github api to sync lightweight tags, but to sync annotated tags we need to use git. Unfortunately there’s no way around that.

pub async fn sync_all_tags( &self, process_annotated: bool, repositories: Vec<String>, ) -> Result<(), GitError>

Sync tags for all configured repositories

pub async fn sync_lightweight_tag( &self, owner: &str, repo: &str, tag: &TagInfo, ) -> Result<(), GitError>

Sync lightweight tags from the parent repo to the forked repo. This can be trivially done using the github api, so we don’t need to call out to

pub async fn sync_annotated_tags( &self, tags: &IndexSet<TagInfo>, ssh_url: &str, parent_urls: HashSet<String>, ) -> Result<(), GitError>

Sync all new annotated tags from a forked repo with its parent. Doing this requires using git (or some re-implementation of git). Syncing annotated tags through the github api with all of its fields, including signing, is currently not possible.

We use tokio::task::spawn_blocking to make sure we don’t make any other async functions hang. Command::new does block.

pub async fn create_tag( &self, url: &str, tag: &str, branch: &str, ) -> Result<(), GitError>

Create a tag for a specific repository

pub async fn create_all_tags( &self, tag: &str, branch: &str, repositories: Vec<String>, ) -> Result<(), GitError>

Create the tag for all configured repositories

pub async fn delete_tag(&self, url: &str, tag: &str) -> Result<(), GitError>

Delete the specified tag for a repository. Deleting a tag does not necessarily return a json response, so we handle this one differently

pub async fn delete_all_tags( &self, tag: &str, repositories: &Vec<String>, ) -> Result<(), GitError>

Delete the specified tag for all configured repositories