git_sync::github

Module match_args

Handle passed arguments

Functions

generate_manpages 🔒

Generate manpages for all subcommands. Otherwise we only get a manpage for the root command.

match_arguments

Parse the argument that gets passed, and run their associated methods

match_backup_cmds 🔒

Process all backup commands

match_branch_cmds 🔒

Process all Branch commands

match_pr_cmds 🔒

Process all PR commands

match_release_cmds 🔒

Process all Release commands

match_repo_cmds 🔒

Process all Repo commands

match_tag_cmds 🔒

Process all Tag commands

write_man 🔒

Helper to write manpages