git_sync

Macro handle_api_response

macro_rules! handle_api_response {
    (
        $response:expr,
        $context:expr,
        $ok:expr,
    ) => { ... };
}

Handles the result of an API call using octocrab and maps them to a GitError. This helps reduce a lot of unnecessary boilerplate. This is used for api calls that don’t already have a high level function provided by Octocrab.

Parameters

$response – The result to match.

$context – A string describing the context of what you’re trying to handle, for logging.

$ok – The closure to run if the result is Ok.

Returns

Returns the result of $ok if successful, or a GitError if the API call fails. You can use ‘?’ to propagate the error to your function.

async fn sync_fork(owner: &str, repo: &str, parent_owner: &str, parent_repo: &str, main_branch: &str) -> Result<(), GitError> {
    let octocrab = octocrab::instance();
    let body = serde_json::json!({"branch": main_branch.to_string()});
    let response: Result<serde_json::Value, octocrab::Error> = octocrab
        .post(format!("/repos/{owner}/{repo}/merge-upstream"), Some(&body))
        .await;

    handle_api_response!(
        response,
        format!("Unable to sync {owner}/{repo} with {parent_owner}/{parent_repo}"),
        |_| {
            println!("Successfully synced {owner}/{repo} with {parent_owner}/{parent_repo}");
            Ok::<(), GitError>(())
        })?;
       Ok(())
   }