git_sync::utils

Module repo

Structs

BranchProtectionRule

Struct to hold branch protection rule information

Checks

A holder for the results of various checks on a repository

LicenseInfo

RepoChecks

A holder for things that can be checked for a repository

RepoInfo

Hold basic information about a github url

TagInfo

Struct for holding tag information. This is for both annotated and lightweight tags.

Enums

TagType

The different types of tags

Statics

REPO_REGEX 🔒

Functions

get_http_status

Convert github http status errors to a usable string message

get_repo_info_from_url

Parse the owner and repository name from a github repository url.

http_to_ssh_repo

Convert an https github url to a clonable SSH URL, needed for cloning a repository in order to be able to push to the repository

opt_bool 🔒